+++ src/js/languages/cs.js
@@ -16,6 +16,9 @@
 		exports.cs = {
 			"mejs.plural-form": 8,
 
+			// core/mediaelement.js
+			"mejs.download-file": "Stáhnout soubor",
+
 			// renderers/flash.js
 			// "mejs.install-flash": "You are using a browser that does not have Flash player enabled or installed. Please turn on your Flash player plugin or download the latest version from https://get.adobe.com/flashplayer/",