+++ changelog.md
@@ -8,6 +8,7 @@
 * Fixed issues in documentation @ron666
 * Fixed issue with `readyState` value in Flash video shim @ron666
 * Added pause event when changing media sources @ron666
+* Integrated `remove()` polyfill to simplify more code (https://github.com/johndyer/mediaelement/pull/2140) @jimmywarting
 
 *3.2.4 (2017/03/14)*